+++ bugsnag/handler.go
@@ -7,6 +7,16 @@ Although that works, makes writing software harder, so this package recommends c
 	APIKey := "key"
 
 	handler := bugsnag.NewHandler(APIKey)
+
+If you need more control over th underlying Notifier instance (eg. more advanced construction),
+you can create a custom one and then create a handler using it:
+
+	// Note: there is a conflict in package names which is resolved here with the "_" (underline) prefix.
+	handler := &_bugsnag.NewHandlerFromNotifier(bugsnag.New(bugsnag.Configuration{
+		APIKey:      APIKey,
+		AppVersion:  "1.0.0",
+		Synchronous: true,
+	}))
 */
 package bugsnag
 
@@ -24,6 +34,11 @@ func NewHandler(APIKey string) *handler {
 			APIKey: APIKey,
 		}),
 	}
+}
+
+// NewHandlerFromNotifier creates a new Bugsnag handler from a notifier instance.
+func NewHandlerFromNotifier(notifier *bugsnag.Notifier) *handler {
+	return &handler{notifier}
 }
 
 // Handle calls the underlying Bugsnag notifier.